Only buffer if node is next leader
After pace maker logic update, when the node becomes the leader it
will immediately broadcast Proposal, thus it should only add the
request to local buffer in case it is the next leader. Otherwise it just
proposed and it may take a while before it becomes leader again and T2
timeout could occur before.